Removes text color transition on input.
This might be temporary, but is the best dark-mode compromise I can come
up with for now. Previously, when the Input was focused, the label
became black, regardless of whether the app was in light or dark mode.
Now it stays gray, which still isn't great but its the best we can get
until I figure out a way to make it black in light mode and white in dark mode.